This always kinda bothered me, Inviting members from a different faction to groups/guilds, now requires GMPermissions rather then just any flag at all. Whispers to GM's now require GMPermissions rather then any flag at all.
This really only affected servers that give out different flags for different things, EX a teleport only flag for vip's, would allow them to invite anyone to their guild/group regardless of faction restriction, and whisper gm's, which can get annoying. If this gets in the way (shouldn't) revert it.

git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@77 a6a5f009-272a-4b40-a74d-5f9816a51f88