Added Pre-Defined kick message support, this is a fun-serverish feature so it is disabled by default. It can be enabled via the ascent-world.conf be sure to update your conf file if you plan to use it.
This will allow you to add/remove predefined kick messages in-game via the AspireMessageHandler, you can use the chat commands aspire addkick and aspire removekick to add/remove a predefined kick message. What this does is lets say you have a common reason for kicking someone, lets say "bug abuse", instead of having to type .kick somenub abusing a bug! you could simply add it as a pre-defined message and type .kick somenub 1 (example) and it would print out the message binded to that index (which you set).

Command usage is, .aspire addkick Index Message, Index being the referce you want to call it, it cannot contain spaces. The message is the server message that will be displayed when this index is called, it can contain spaces. 
.aspire removekick Index will remove the pre-defined kick message binded to that index.

It will store these messages in the database, Requires server_kick_messages.sql. Consider this NTY for now, have not stability tested it.

git-svn-id: svn://svn.aspiredev.org/AspireCore/trunk@424 a6a5f009-272a-4b40-a74d-5f9816a51f88